requirements.txt
amqp
aniso8601
appdirs
asgiref
billiard
CacheControl
celery
certifi
chardet
click
click-didyoumean
click-plugins
click-repl
colorama
contextlib2
distlib
distro
Django
django-cors-headers
django-filter
django-graphql-auth
django-graphql-jwt
django-session-security
django-widget-tweaks
filelock
future
graphene
graphene-django
graphql-core
graphql-relay
html5lib
idna
ipaddr
kombu
lockfile
msgpack
packaging
pep517
Pillow
progress
promise
prompt-toolkit
psycopg2-binary
PyJWT
pyparsing
pyserial
pytoml
pytz
redis
requests
requests-file
retrying
Rx
selenium
singledispatch
six
sqlparse
tablib
text-unidecode
tldextract
urllib3
vine
wcwidth
webencodings
XlsxWriter